src/app/components/to-do-list/to-do-list.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { ToDoListComponent } from './to-do-list.component';

describe('ToDoListComponent', () => {
  let component: ToDoListComponent;
  let fixture: ComponentFixture<ToDoListComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ToDoListComponent],
    }).compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(ToDoListComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should add a new task to the todolist', () => {
    expect(component.todolist.length).toEqual(component.todolist.length);

    component.addNewTask(component.newTask, component.newTaskTime);

    expect(component.todolist.length).toEqual(component.todolist.length++);
  });
});
